On CI environments run createcoverage without opening a browser. (#24)
Also, we were previously calling 'bin/createcoverage run bin/test'.
The 'run bin/test' part is useless, so I removed it.  Maybe it was
needed in previous versions.

We check the $CI environment variable to see if we are on a Continuous
Integration machine.  This at least works for Travis:
https://docs.travis-ci.com/user/environment-variables#Default-Environment-Variables

In that case, run bin/createcoverage with the `--output-dir` option.
This has the side effect of not opening a browser to view the results.
This has two reasons:
1. There is no user who can look at the results, so it is useless.
2. Travis fails to kill the browser at the end, so the build never finishes.
See https://github.com/reinout/createcoverage/issues/5

Sample failing build:
https://travis-ci.org/collective/collective.watcherlist/jobs/185787861